+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 # MyToDo changelog
+# 27/06/2021 (Version 1.2)
+### New
+- Add list to-dos capability
+### Changes
+- List the groups more efficiently, using indexes and normalizing the noSQL model
+- Now edit the groups by clicking in an edit icon next to the group name
+### Fixes
+- Show the message when the groups list is empty
 
 # 25/06/2021 (Version 1.1)
 ### New